+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## vNEXT
 - Match orders boost through voucher. (#16)
+- Use hardhat deploy. (#15)
 - Upload coverage reports to Codecov. (#14)
 - Clean some TODOs. (#13)
 - Match orders through voucher. (#12)